refactor: remove obsolete snapshot-writing APIs
