+++ packages/assistant-runtime/test/package-entrypoints.test.ts
@@ -22,7 +22,6 @@ import {
   runHostedWorkspaceInvocation as runHostedWorkspaceInvocationPublic,
 } from "@murphai/assistant-runtime/hosted-invocation";
 import {
-  checkpointHostedRuntimeBridgeWorkspace as checkpointHostedRuntimeBridgeWorkspaceTestkitPublic,
   createHostedWorkspaceRuntimeBridgeJobOptions as createHostedWorkspaceRuntimeBridgeJobOptionsPublic,
 } from "@murphai/assistant-runtime/hosted-invocation-testkit";
 import {
@@ -66,7 +65,6 @@ import {
   runHostedWorkspaceInvocation as runHostedWorkspaceInvocationDirect,
 } from "../src/hosted-invocation.ts";
 import {
-  checkpointHostedRuntimeBridgeWorkspace as checkpointHostedRuntimeBridgeWorkspaceTestkitDirect,
   createHostedWorkspaceRuntimeBridgeJobOptions as createHostedWorkspaceRuntimeBridgeJobOptionsDirect,
 } from "../src/hosted-invocation-testkit.ts";
 import {
@@ -269,10 +267,6 @@ test("hosted-invocation-testkit subpath export stays wired to bridge option cons
   assert.equal(
     createHostedWorkspaceRuntimeBridgeJobOptionsPublic,
     createHostedWorkspaceRuntimeBridgeJobOptionsDirect,
-  );
-  assert.equal(
-    checkpointHostedRuntimeBridgeWorkspaceTestkitPublic,
-    checkpointHostedRuntimeBridgeWorkspaceTestkitDirect,
   );
 });
